+++ app/Console/Kernel.php
@@ -36,7 +36,7 @@ class Kernel extends ConsoleKernel
         $schedule->call(function(){$this->diario01();})->dailyAt('06:00')->timezone(Config::get('constants.USO_HORARIO_ARG'));
         $schedule->call(function(){$this->cadaMinuto();})->everyMinute();
         //$schedule->call(function(){$this->cadaCincoMinutos();})->everyFiveMinutes();
-        //$schedule->call(function(){$this->mensual();})->monthly()->timezone(Config::get('constants.USO_HORARIO_ARG'));
+        $schedule->call(function(){$this->mensual();})->monthly()->timezone(Config::get('constants.USO_HORARIO_ARG'));
         ### ->monthly(); //Run the task on the first day of every month at 00:00
     }
 
@@ -64,7 +64,7 @@ class Kernel extends ConsoleKernel
 
     private function mensual ()
     {
-        //CronFunciones::resetCounter(true);
+        CronFunciones::resetContadores_mensuales();
     }
     private function semanal ()
     {